requirements.txt
click~=7.0
requests~=2.21.0
requirements-parser~=0.2.0
readsettings~=3.2.0
